[ProcessContainers] C interface for containers
