+++ Source/processcontainers/CMakeLists.txt
@@ -3,7 +3,7 @@ set(TARGET  ${NAMESPACE}ProcessContainers)
 # Construct a library object
 add_library(${TARGET} SHARED
         ProcessContainer.cpp
-        LXCImplementation.cpp
+        implementations/LXCImplementation/LXCImplementation.cpp
         Module.cpp
         )
